A bunch of updates to work better with Game.Default and Clock.
This includes some breaking API changes, for which I apologize, but things were so wonky inside TP
that I figured I'd better just break things now to bring them in line than try to provide
half-assed deprecated migration methods.

Basically, you'll need to create a Clock in your game if you're going to use TP and pass it into
the places that need passing. The API is much more self-explanatory now. If there's an update(int)
method, you should call it from your Game.Default.update(int) method. If there's a paint(Clock)
method, you should call it from your Game.Default.paint(float) method, after having updated your
clock.

Previously there were methods named update() that were supposed to be called from paint() and a
bunch of places where things that should have been run on paint() using an alpha-adjust time were
run on update() using simulation time. This resulted in animations that looked crappy unless you
cranked up your game's simulation update time, which is a bad idea. Now you can use a sensible
update time, like 50ms, and animations remain smooth (are in fact smoother, since things are
properly interpolated frame-by-frame).

There are a few places where more work is needed (Flump for example), but I wanted to get this big
smack in the pants out of the way before digging into the niggly bits.

+++ core/src/main/java/tripleplay/particle/Particles.java
@@ -49,7 +49,7 @@ public class Particles
      *
      * @param delta the time that has elapsed since the last frame, in milliseconds.
      */
-    public void update (float delta) {
+    public void update (int delta) {
         _onUpdate.emit(_now.update(delta));
     }
 
@@ -67,8 +67,8 @@ public class Particles
     protected static class Now {
         public float time;
         public float dt;
-        public Now update (float delta) {
-            dt = delta/1000;
+        public Now update (int delta) {
+            dt = delta/1000f;
             time += dt;
             return this;
         }
